+++ test/unit/gradient.js
@@ -666,49 +666,43 @@
   });
 
   QUnit.test('toSVG linear', function(assert) {
-    fabric.Object.__uid = 0;
     var gradient = createLinearGradient();
     var obj = new fabric.Object({ width: 100, height: 100 });
-    assert.equal(gradient.toSVG(obj), SVG_LINEAR);
+    assert.equalSVG(gradient.toSVG(obj), SVG_LINEAR);
   });
 
   QUnit.test('toSVG radial', function(assert) {
-    fabric.Object.__uid = 0;
     var gradient = createRadialGradient();
     var obj = new fabric.Object({ width: 100, height: 100 });
-    assert.equal(gradient.toSVG(obj), SVG_RADIAL);
+    assert.equalSVG(gradient.toSVG(obj), SVG_RADIAL);
   });
 
   QUnit.test('toSVG radial with r1 > 0', function(assert) {
-    fabric.Object.__uid = 0;
     var gradient = createRadialGradientWithInternalRadius();
     var obj = new fabric.Object({ width: 100, height: 100 });
-    assert.equal(gradient.toSVG(obj), SVG_INTERNALRADIUS);
+    assert.equalSVG(gradient.toSVG(obj), SVG_INTERNALRADIUS);
   });
 
   QUnit.test('toSVG radial with r1 > 0 swapped', function(assert) {
-    fabric.Object.__uid = 0;
     var gradient = createRadialGradientSwapped();
     var obj = new fabric.Object({ width: 100, height: 100 });
     const gradientColorStops = JSON.stringify(gradient.colorStops);
-    assert.equal(gradient.toSVG(obj), SVG_SWAPPED, 'it exports as expected');
+    assert.equalSVG(gradient.toSVG(obj), SVG_SWAPPED, 'it exports as expected');
     const gradientColorStopsAfterExport = JSON.stringify(gradient.colorStops);
-    assert.equal(gradient.toSVG(obj), SVG_SWAPPED, 'it exports as expected a second time');
-    assert.equal(gradientColorStops, gradientColorStopsAfterExport, 'colorstops do not change')
+    assert.equalSVG(gradient.toSVG(obj), SVG_SWAPPED, 'it exports as expected a second time');
+    assert.equalSVG(gradientColorStops, gradientColorStopsAfterExport, 'colorstops do not change')
   });
 
   QUnit.test('toSVG linear objectBoundingBox', function(assert) {
-    fabric.Object.__uid = 0;
     var gradient = createLinearGradient('percentage');
     var obj = new fabric.Object({ width: 100, height: 100 });
-    assert.equal(gradient.toSVG(obj), SVG_LINEAR_PERCENTAGE);
+    assert.equalSVG(gradient.toSVG(obj), SVG_LINEAR_PERCENTAGE);
   });
 
   QUnit.test('toSVG radial objectBoundingBox', function(assert) {
-    fabric.Object.__uid = 0;
     var gradient = createRadialGradient('percentage');
     var obj = new fabric.Object({ width: 100, height: 100 });
-    assert.equal(gradient.toSVG(obj), SVG_RADIAL_PERCENTAGE);
+    assert.equalSVG(gradient.toSVG(obj), SVG_RADIAL_PERCENTAGE);
   });
 
 })();